[refactor] 파일 도메인 수정
