+++ templates/JS/register.js
